+++ src/components/modules/Lesson.vue
@@ -8,6 +8,8 @@
               <q-item link v-close-overlay @click.native="edit(id)">Design</q-item>
               <q-item link v-close-overlay @click.native="$router.push({ name: 'rlesson', params: { seriesid: $route.params.seriesid, lessonid: id } })">Devos</q-item>
               <q-item link v-close-overlay @click.native="$router.push({ name: 'rresearch', params: { seriesid: $route.params.seriesid, lessonid: id } })">Research</q-item>
+              <q-item link v-close-overlay @click.native="$router.push({ name: 'rguide', params: { seriesid: $route.params.seriesid, lessonid: id } })">Guides</q-item>
+              <q-item link v-close-overlay @click.native="$router.push({ name: 'rreview', params: { seriesid: $route.params.seriesid, lessonid: id } })">Review</q-item>
               <q-item link @click.native="remove(id)">Delete</q-item>
             </q-list>
           </q-popover>